+++ fdb-record-layer-lucene/src/main/java/com/apple/foundationdb/record/lucene/codec/LuceneOptimizedCompoundFormat.java
@@ -49,10 +49,10 @@ public class LuceneOptimizedCompoundFormat extends CompoundFormat {
     public static final int VERSION_START = 0;
     static final int VERSION_CURRENT = VERSION_START;
 
-    private final CompoundFormat compoundFormat;
+    protected final CompoundFormat underlying;
 
     public LuceneOptimizedCompoundFormat(final CompoundFormat underlying) {
-        this.compoundFormat = underlying;
+        this.underlying = underlying;
     }
 
     @Override
@@ -75,8 +75,12 @@ public class LuceneOptimizedCompoundFormat extends CompoundFormat {
         si.setFiles(filteredFiles);
         @SuppressWarnings("PMD.CloseResource") // we don't need to close this because it is just extracting from the dir
         final FDBDirectory directory = FDBDirectoryUtils.getFDBDirectoryNotCompound(dir);
-        compoundFormat.write(dir, si, context);
+        underlying.write(dir, si, context);
         si.setFiles(filesForAfter);
+        copyFieldInfos(si, filesForAfter, directory);
+    }
+
+    protected void copyFieldInfos(final SegmentInfo si, final Set<String> filesForAfter, final FDBDirectory directory) {
         final String fieldInfosFileName = filesForAfter.stream().filter(FDBDirectory::isFieldInfoFile).findFirst().orElseThrow();
         final FDBLuceneFileReference fieldInfosReference = directory.getFDBLuceneFileReference(fieldInfosFileName);
         String entriesFile = IndexFileNames.segmentFileName(si.name, "", ENTRIES_EXTENSION);
@@ -100,6 +104,12 @@ public class LuceneOptimizedCompoundFormat extends CompoundFormat {
             }
         }
 
+        validateFileCounts(files, fieldInfos, storedFields);
+
+        return filteredFiles;
+    }
+
+    protected void validateFileCounts(final Set<String> files, final int fieldInfos, final int storedFields) {
         if (fieldInfos != 1) {
             throw new RecordCoreException("Segment has wrong number of FieldInfos")
                     .addLogInfo(LuceneLogMessageKeys.FILE_LIST, files);